chore: export core-logger-winston types.
